* fix(1234-Windows-port-new): fix windows port.

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix FTP passive port

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix dynamodb container creation

* fix(1234-Windows-port-new): fix application startup

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix application startup

* fix(1234-Windows-port-new): fix s3 handler stol issue

* Merge branch 'main' into fix/1234-Windows-port-new

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix s3 handler stol issue

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add sorting to auto load files in directory

* Merge branch 'main' into fix/1234-Windows-port-new

* Merge branch 'main' into fix/1233-Windows-port-new

* fix(1234-Windows-port-new): add application watchdog

* fix(1234-Windows-port-new): add application watchdog

* fix(1234-Windows-port-new): add application watchdog

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add application watchdog

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): new WindowsService.cpp

* fix(1234-Windows-port-new): new WindowsService.cpp

* fix(1234-Windows-port-new): new WindowsService.cpp

* fix(1234-Windows-port-new): add application log server port

* Merge remote-tracking branch 'origin/main' into fix/1234-Windows-port…

* fix(1234-Windows-port-new): add application log server port

* Merge branch 'main' into fix/1234-Windows-port-new

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): add application log server port

* fix(1234-Windows-port-new): add application log server port

* fix(1234-Windows-port-new): add application log server port

* fix(1234-Windows-port-new): add application log server port

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix application watchdog

* fix(1234-Windows-port-new): fix application watchdog

* Merge remote-tracking branch 'origin/main' into fix/1234-Windows-port…

* fix(1234-Windows-port-new): fix application watchdog

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix windows port

* fix(1234-Windows-port-new): fix windows port

* Merge remote-tracking branch 'origin/main' into fix/1234-Windows-port…

* fix(1234-Windows-port-new): fix windows port

* Merge branch 'main' into fix/1234-Windows-port-new

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): stop will disable application

* fix(1234-Windows-port-new): add FTP  cognito connection

* fix(1234-Windows-port-new): fix cognito FE pages

* Merge remote-tracking branch 'origin/main' into fix/1234-Windows-port…

* fix(1234-Windows-port-new): fix cognito FE pages

* Merge remote-tracking branch 'origin/main' into fix/1234-Windows-port…

+++ src/dto/include/awsmock/dto/sns/internal/ListMessageCountersResponse.h
@@ -25,15 +25,15 @@ namespace AwsMock::Dto::SNS {
          */
         long total = 0;
 
-      private:
+    private:
 
         friend ListMessageCountersResponse tag_invoke(boost::json::value_to_tag<ListMessageCountersResponse>, boost::json::value const &v) {
             ListMessageCountersResponse r;
             r.region = v.at("region").as_string();
             r.user = v.at("user").as_string();
             r.requestId = v.at("requestId").as_string();
-            r.total = v.at("total").as_int64();
-            r.messages = boost::json::value_to<std::vector<MessageCounter>>(v.at("messages"));
+            r.total = static_cast<long>(v.at("total").as_int64());
+            r.messages = boost::json::value_to<std::vector<MessageCounter> >(v.at("messages"));
             return r;
         }
